SCE-504: Moved launcher interfaces to executor pkg
It is needed since we will use these interfaces in different packages than workloads.

- Moved launcher interfaces to executor.


# Conflicts:
#	pkg/workloads/caffe/caffe.go
