+++ pkg/executor/load_generator.go
@@ -1,7 +1,6 @@
-package workloads
+package executor
 
 import (
-	"github.com/intelsdi-x/swan/pkg/executor"
 	"time"
 )
 
@@ -17,5 +16,5 @@ type LoadGenerator interface {
 	// Load starts a load on the specific workload with the defined loadPoint (number of QPS).
 	// The task will do the load for specified amount of time.
 	// Note: Results from Load needs to be fetched out of band e.g using Snap.
-	Load(load int, duration time.Duration) (task executor.TaskHandle, err error)
+	Load(load int, duration time.Duration) (task TaskHandle, err error)
 }
