+++ pkg/workloads/low_level/stream/stream.go
@@ -7,7 +7,6 @@ import (
 	"github.com/intelsdi-x/swan/pkg/conf"
 	"github.com/intelsdi-x/swan/pkg/executor"
 	"github.com/intelsdi-x/swan/pkg/utils/fs"
-	"github.com/intelsdi-x/swan/pkg/workloads"
 )
 
 const (
@@ -53,7 +52,7 @@ type stream struct {
 // Working set size should be more than 4x the size of sum of all last-level cache used in the run
 // If you need more consider rebuilding stream with STREAM_ARRAY_SIZE adjusted accordingly.
 // Check stream.c "Instructions" for more details.
-func New(exec executor.Executor, config Config) workloads.Launcher {
+func New(exec executor.Executor, config Config) executor.Launcher {
 	return stream{
 		exec: exec,
 		conf: config,
